// src/lib/types.ts
export type EventCategory =
  | "SEMINAR_WORKSHOP"
  | "RAPAT_KOORDINASI"
  | "DEADLINE_PROYEK"
  | "KEGIATAN_SOSIAL"
  | "KOMPETISI_LOMBA"
  | "INTERNAL_TEAM"
  | string;

export type EventPriority = "LOW" | "MEDIUM" | "HIGH" | "URGENT";

export type EventStatus =
  | "UPCOMING"
  | "IN_PROGRESS"
  | "COMPLETED"
  | "POSTPONED"
  | "CANCELLED";

export type LocationType = "OFFLINE" | "ONLINE" | "HYBRID";

export type TaskStatus = "TODO" | "IN_PROGRESS" | "REVIEW" | "DONE";